refactor: Phase 12–13a — transport split, App.vue decomposition into 5 composables
Phase 12: Split llmApi.js into transport modules (requestOrchestrator,
reasoningHeaders), remove dead params from streamAccumulator/streamingSse.

Phase 13a: Extract App.vue (1229 → 622 lines) into composables:
- useAppNavigation.js (101 lines) — view routing, desktop/mobile
- useEditorController.js (304 lines) — editor lifecycle, configs, FS editor
- useGlossaryPopup.js (99 lines) — desktop glossary drag popup
- useAppEventSubscriptions.js (190 lines) — 25+ event subs, cleanup
- useAppInit.js (108 lines) — onMounted init, onBeforeUnmount cleanup
