+++ src/App.vue
@@ -1,5 +1,5 @@
 <script setup>
-import { ref, reactive, onMounted, nextTick, computed, watch, onBeforeUnmount, defineAsyncComponent } from 'vue';
+import { ref, reactive, computed, watch, nextTick, defineAsyncComponent } from 'vue';
 import AppHeader from '@/components/layout/AppHeader.vue';
 import BottomNavigation from '@/components/layout/BottomNavigation.vue';
 import DialogList from '@/views/DialogList.vue';
@@ -19,7 +19,6 @@ const ApiView = defineAsyncComponent(() => import('@/views/ApiView.vue'));
 const Editor = defineAsyncComponent(() => import('@/components/editors/GenericEditor.vue'));
 const FullScreenEditor = defineAsyncComponent(() => import('@/components/editors/FullScreenEditor.vue'));
 
-// const HoloCardViewer = defineAsyncComponent(() => import('@/components/media/HoloCardViewer.vue'));
 const ImageViewer = defineAsyncComponent(() => import('@/components/media/ImageViewer.vue'));
 import AppToast from '@/components/ui/AppToast.vue';
 import MagicDrawer from '@/components/chat/MagicDrawer.vue';
@@ -37,41 +36,26 @@ const DragDropOverlay = defineAsyncComponent(() => import('@/components/ui/DragD
 const ToolsView = defineAsyncComponent(() => import('@/views/ToolsView.vue'));
 const PersonasView = defineAsyncComponent(() => import('@/views/PersonasView.vue'));
 const RegexSheet = defineAsyncComponent(() => import('@/components/sheets/RegexSheet.vue'));
-import { Capacitor } from '@capacitor/core';
-import { isKeyboardOpen, onKeyboardShow, onKeyboardHide } from '@/core/services/keyboardHandler.js';
-import { initSettings, applyApiRuntimeConfig } from '@/core/config/APISettings.js';
-import { initTheme, themeState } from '@/core/states/themeState.js';
-import { updateLanguage, translations } from '@/utils/i18n.js';
-import { currentLang, imageViewerMode, forceMobileLayout } from '@/core/config/APPSettings.js';
-import { initRipple, initThemeToggle, initHeaderDropdown, initBackButton, initViewportFix } from '@/core/services/ui.js';
-import { bottomSheetState, closeBottomSheet, showBottomSheet } from '@/core/states/bottomSheetState.js';
-import { db, migrateScToGz, markSyncDeletedEntry } from '@/utils/db.js';
-import { addPersona, updatePersona, deletePersona, allPersonas, loadPersonas } from '@/core/states/personaState.js';
-import { checkAndRequestNotifications, consumePendingNotificationData } from '@/core/services/notificationService.js';
-import { logger } from './utils/logger.js';
-import { generateMissingThumbnails } from '@/utils/characterIO.js';
-import { initLorebookState } from '@/core/states/lorebookState.js';
-import { initPresetState } from '@/core/states/presetState.js';
+
+import { isKeyboardOpen } from '@/core/services/keyboardHandler.js';
+import { themeState } from '@/core/states/themeState.js';
+import { translations } from '@/utils/i18n.js';
+import { currentLang } from '@/core/config/APPSettings.js';
+import { bottomSheetState, closeBottomSheet } from '@/core/states/bottomSheetState.js';
 import { sidebarState } from '@/core/states/sidebarState.js';
-import { initSyncState, syncProvider } from '@/core/states/syncState.js';
-import { fullPull, checkSyncReadiness } from '@/core/services/syncService.js';
-import { startTracking } from '@/core/services/timeTracker.js';
 import { APP_EVENTS } from '@/core/events/eventNames.js';
-import { publishAppEvent, subscribeAppEvent } from '@/core/events/eventHub.js';
+import { publishAppEvent } from '@/core/events/eventHub.js';
 
+import { useAppNavigation } from '@/composables/app/useAppNavigation.js';
+import { useEditorController, characterEditorConfig, personaEditorConfig } from '@/composables/app/useEditorController.js';
+import { useGlossaryPopup } from '@/composables/app/useGlossaryPopup.js';
+import { useAppEventSubscriptions } from '@/composables/app/useAppEventSubscriptions.js';
+import { useAppInit } from '@/composables/app/useAppInit.js';
 
-
-const t = (key) => translations[currentLang.value]?.[key] || key;
-// Initialize error handling
-
-// --- Navigation state ---
-const isDesktop = ref(typeof window !== 'undefined' && window.innerWidth >= 768 && !forceMobileLayout.value);
-const currentView = ref(isDesktop.value ? 'view-characters' : 'view-dialogs');
-const headerRef = ref(null); // Reference to the AppHeader component
+// --- Template refs ---
+const headerRef = ref(null);
 const headerContainer = ref(null);
 const footerContainer = ref(null);
-const isChatInitialized = ref(false); // Whether the chat has been initialized
-const isDataLoaded = ref(false);
 const dialogListRef = ref(null);
 const characterListRef = ref(null);
 const chatViewRef = ref(null);
@@ -82,124 +66,9 @@ const syncSheetRef = ref(null);
 const conflictSheetRef = ref(null);
 const presetViewRef = ref(null);
 const apiViewRef = ref(null);
+const isDataLoaded = ref(false);
 
-const isHeaderEditorMode = ref(false);
-
-const checkDesktop = () => { 
-    const isDesk = typeof window !== 'undefined' && window.innerWidth >= 768 && !forceMobileLayout.value;
-    if (isDesk !== isDesktop.value) {
-        if (isDesk && currentView.value === 'view-dialogs') {
-            currentView.value = 'view-characters';
-        }
-        isDesktop.value = isDesk;
-    }
-};
-
-const menuViews = ['view-menu', 'view-settings', 'view-theme-settings'];
-const isDesktopFloating = computed(() => isDesktop.value && menuViews.includes(currentView.value));
-
-const isGlossaryWindowOpen = ref(false);
-const isDesktopGlossary = computed(() => isDesktop.value && isGlossaryWindowOpen.value);
-const glossaryPopupTitle = ref('');
-const glossaryCanGoBack = ref(false);
-
-function onGlossaryHeaderUpdate(detail) {
-    if (detail?.title !== undefined) glossaryPopupTitle.value = detail.title;
-    if (detail?.canGoBack !== undefined) glossaryCanGoBack.value = detail.canGoBack;
-}
-function onGlossaryBack() {
-    publishAppEvent(APP_EVENTS.ui.glossary.back);
-}
-
-const glossaryPos = reactive({ x: 0, y: 0 });
-const glossaryStyle = computed(() => {
-    if (glossaryPos.x === 0 && glossaryPos.y === 0) return {};
-    return {
-        top: `${glossaryPos.y}px`,
-        left: `${glossaryPos.x}px`,
-        bottom: 'auto',
-        right: 'auto',
-        transform: 'none',
-        margin: 0
-    };
-});
-
-let isDraggingGlossary = false;
-let glossaryDragStartX = 0, glossaryDragStartY = 0;
-let glossaryInitialX = 0, glossaryInitialY = 0;
-const appEventUnsubs = [];
-
-function startGlossaryDrag(e) {
-    if (e.target.closest('button')) return;
-    isDraggingGlossary = true;
-    glossaryDragStartX = e.clientX;
-    glossaryDragStartY = e.clientY;
-    
-    const popup = document.querySelector('.desktop-glossary-popup');
-    if (!popup) return;
-    const rect = popup.getBoundingClientRect();
-    if (glossaryPos.x === 0 && glossaryPos.y === 0) {
-        glossaryPos.x = rect.left;
-        glossaryPos.y = rect.top;
-    }
-    glossaryInitialX = glossaryPos.x;
-    glossaryInitialY = glossaryPos.y;
-    
-    window.addEventListener('mousemove', onGlossaryDrag);
-    window.addEventListener('mouseup', stopGlossaryDrag);
-}
-
-function onGlossaryDrag(e) {
-    if (!isDraggingGlossary) return;
-    const dx = e.clientX - glossaryDragStartX;
-    const dy = e.clientY - glossaryDragStartY;
-    glossaryPos.x = Math.max(0, Math.min(window.innerWidth - 100, glossaryInitialX + dx));
-    glossaryPos.y = Math.max(0, Math.min(window.innerHeight - 50, glossaryInitialY + dy));
-}
-
-function stopGlossaryDrag() {
-    isDraggingGlossary = false;
-    window.removeEventListener('mousemove', onGlossaryDrag);
-    window.removeEventListener('mouseup', stopGlossaryDrag);
-}
-
-const desktopPreMenuView = ref(isDesktop.value ? 'view-characters' : 'view-dialogs');
-
-const effectiveMainView = computed(() => {
-    if (isDesktopFloating.value) return desktopPreMenuView.value;
-    return currentView.value;
-});
-
-const waitForComponent = (refVar, callback) => {
-    if (refVar.value) {
-        callback(refVar.value);
-    } else {
-        const unwatch = watch(refVar, (val) => {
-            if (val) {
-                callback(val);
-                unwatch();
-            }
-        });
-    }
-};
-
-// --- Editor State ---
-const editingCharacter = ref(null);
-const editingCharacterIndex = ref(-1);
-const editingPersona = ref(null);
-const editingPersonaIndex = ref(-1);
-const previousViewForEditor = ref(null);
-const previousSessionIdForEditor = ref(null);
-const currentChatSessionId = ref(null);
-const activeChatCharObj = ref(null);
-const chatPreviousView = ref(null);
-const shouldOpenPersonasOnReturn = ref(false);
-
-const isDeleting = ref(false); // Guard flag to prevent auto-save during deletion
-const isOnboarding = ref(false);
-const kbListeners = [];
-
-// --- Categories ---
+// --- Categories (must be before useAppInit which uses them) ---
 const activeCategories = reactive({
     'view-dialogs': 'all',
     'view-characters': 'all'
@@ -218,257 +87,131 @@ const categories = {
     ]
 };
 
-const isEditorView = computed(() => currentView.value === 'view-character-edit' || currentView.value === 'view-persona-edit');
+// --- Composables ---
+const nav = useAppNavigation();
+const editor = useEditorController({
+    currentView: nav.currentView,
+    currentChatSessionId: nav.currentChatSessionId,
+    waitForComponent,
+    chatViewRef
+});
+const glossary = useGlossaryPopup({ isDesktop: nav.isDesktop });
 
-const fsEditorVisible = ref(false);
-const headerZIndex = computed(() => {
-    if (fsEditorVisible.value) return 2001;
-    if (isEditorView.value) return 1100;
-    return 100;
+const events = useAppEventSubscriptions({
+    isDesktop: nav.isDesktop,
+    currentView: nav.currentView,
+    isHeaderEditorMode: nav.isHeaderEditorMode,
+    activeChatCharObj: nav.activeChatCharObj,
+    currentChatSessionId: nav.currentChatSessionId,
+    chatPreviousView: nav.chatPreviousView,
+    isGlossaryWindowOpen: glossary.isGlossaryWindowOpen,
+    openCharacterEditor: editor.openCharacterEditor,
+    onOpenPersonaEditor: editor.onOpenPersonaEditor,
+    openChatWrapper: (char) => nav.openChatWrapper(char, { chatViewRef, waitForComponent }),
+    openFsEditor: editor.openFsEditor,
+    waitForComponent,
+    chatViewRef,
+    connectionsSheetRef,
+    lorebookSheetRef,
+    backupSheetRef,
+    syncSheetRef,
+    conflictSheetRef,
+    presetViewRef,
+    apiViewRef,
+    characterListRef,
+    headerRef,
+    fsEditorVisible: editor.fsEditorVisible,
+    closeAndSaveFsEditor: editor.closeAndSaveFsEditor,
+    handleGlossaryOpen: glossary.handleGlossaryOpen,
+    handleGlossaryToggle: glossary.handleGlossaryToggle,
+    onGlossaryHeaderUpdate: glossary.onGlossaryHeaderUpdate,
+    onLanguageChanged,
+    finishOnboarding: nav.finishOnboarding
 });
 
-const mainZIndex = computed(() => {
-    if (isEditorView.value) return 1000;
-    return 1;
+useAppInit({
+    isOnboarding: nav.isOnboarding,
+    isDataLoaded,
+    isDesktop: nav.isDesktop,
+    checkDesktop: nav.checkDesktop,
+    updateLayoutMetrics: nav.updateLayoutMetrics,
+    headerContainer,
+    footerContainer,
+    categories,
+    activeCategories,
+    appEventUnsubs: events.appEventUnsubs,
+    handleOpenChatEvent: events.handleOpenChatEvent
 });
 
-function openChatWrapper(char) {
-    currentChatSessionId.value = char.sessionId || null;
-    activeChatCharObj.value = char;
-    const previousView = currentView.value;
-    chatPreviousView.value = previousView;
-    currentView.value = 'view-chat';
-    isChatInitialized.value = true;
-    
-    waitForComponent(chatViewRef, (comp) => {
-        comp.openChat(char, () => {
-            currentView.value = previousView;
-            // DialogList updates automatically via events
+events.registerAll();
+
+// --- Local helpers ---
+function waitForComponent(refVar, callback) {
+    if (refVar.value) {
+        callback(refVar.value);
+    } else {
+        const unwatch = watch(refVar, (val) => {
+            if (val) {
+                callback(val);
+                unwatch();
+            }
         });
-    });
+    }
+}
+
+function openChatFromTemplate(char) {
+    nav.openChatWrapper(char, { chatViewRef, waitForComponent });
 }
 
 function onLanguageChanged() {
     if (headerRef.value) headerRef.value.updateHeader();
 }
 
-// --- Editor Logic ---
-async function openCharacterEditor(index) {
-    previousViewForEditor.value = currentView.value;
-    if (currentView.value === 'view-chat') {
-        previousSessionIdForEditor.value = currentChatSessionId.value;
-    } else {
-        previousSessionIdForEditor.value = null;
+// --- Z-index computed ---
+const headerZIndex = computed(() => {
+    if (editor.fsEditorVisible.value) return 2001;
+    if (editor.isEditorView.value) return 1100;
+    return 100;
+});
+
+const mainZIndex = computed(() => {
+    if (editor.isEditorView.value) return 1000;
+    return 1;
+});
+
+// --- Layout metrics ---
+nav.setLayoutMetricsUpdater(() => {
+    if (headerContainer.value) {
+        const h = headerContainer.value.offsetHeight;
+        document.documentElement.style.setProperty('--header-height', `${h}px`);
     }
-    isDeleting.value = false;
-    editingCharacterIndex.value = index;
-    if (index === -1) {
-        editingCharacter.value = null; // New character
-    } else {
-        const chars = (await db.getAll('characters')) || [];
-        editingCharacter.value = chars[index];
+    if (nav.isDesktop.value) {
+        document.documentElement.style.setProperty('--footer-height', '0px');
+    } else if (footerContainer.value) {
+        document.documentElement.style.setProperty('--footer-height', `${footerContainer.value.offsetHeight}px`);
     }
-    currentView.value = 'view-character-edit';
-}
+});
 
-async function handleHeaderSave() {
-    if (currentView.value === 'view-character-edit') {
-        if (editingCharacter.value && editingCharacter.value.name && editingCharacter.value.name.trim() !== '') {
-            await db.saveCharacter(editingCharacter.value, editingCharacterIndex.value);
-            closeEditor();
-        } else {
-            showBottomSheet({
-                title: t('title_error') || 'Error',
-                bigInfo: {
-                    icon: '<svg viewBox="0 0 24 24" style="fill:#ff4444;width:100%;height:100%;"><path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm1 15h-2v-2h2v2zm0-4h-2V7h2v6z"/></svg>',
-                    description: t('error_name_required') || 'Name is required.',
-                    buttonText: t('btn_ok') || 'OK',
-                    onButtonClick: closeBottomSheet
-                }
-            });
-        }
-    } else if (currentView.value === 'view-persona-edit') {
-        if (editingPersona.value && editingPersona.value.name && editingPersona.value.name.trim() !== '') {
-            if (editingPersonaIndex.value === -1) {
-                await addPersona(editingPersona.value);
-            } else {
-                await updatePersona(editingPersonaIndex.value, editingPersona.value);
-            }
-            closeEditor();
-        } else {
-            showBottomSheet({
-                title: t('title_error') || 'Error',
-                bigInfo: {
-                    icon: '<svg viewBox="0 0 24 24" style="fill:#ff4444;width:100%;height:100%;"><path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm1 15h-2v-2h2v2zm0-4h-2V7h2v6z"/></svg>',
-                    description: t('error_name_required') || 'Name is required.',
-                    buttonText: t('btn_ok') || 'OK',
-                    onButtonClick: closeBottomSheet
-                }
-            });
-        }
+// --- FAB Logic ---
+const fabConfig = computed(() => {
+    if (nav.currentView.value === 'view-dialogs' && !nav.isDesktop.value) {
+        return {
+            text: translations[currentLang.value]?.btn_new_chat || 'New Chat',
+            action: () => dialogListRef.value?.openNewChatPicker()
+        };
+    } else if (nav.currentView.value === 'view-characters' && !nav.isDesktop.value) {
+        return {
+            text: translations[currentLang.value]?.btn_add || 'Add',
+            action: () => characterListRef.value?.onAddCharacter()
+        };
     }
-}
+    return null;
+});
 
-async function handleEditorAutoSave(val) {
-    // CRITICAL: Do not save if we are in the process of deleting
-    if (isDeleting.value) return;
-
-    if (currentView.value === 'view-character-edit') {
-        if (!val || !val.name || val.name.trim() === '') return; // Don't auto-save character without name
-
-        if (editingCharacterIndex.value === -1) {
-            // Create new character
-            await db.saveCharacter(val, -1);
-            // Retrieve new index (assuming added to end) to prevent duplicates on next save
-            const chars = (await db.getAll('characters')) || [];
-            editingCharacterIndex.value = chars.length - 1;
-        } else {
-            await db.saveCharacter(val, editingCharacterIndex.value);
-        }
-    } else if (currentView.value === 'view-persona-edit') {
-        if (!val.name) return; // Don't save empty personas on exit
-
-        if (editingPersonaIndex.value === -1) {
-            // Create new persona
-            const newPersona = await addPersona(val);
-            // Update index to point to the new persona (last one)
-            editingPersonaIndex.value = allPersonas.value.length - 1;
-            // Update the editing object to include the generated ID to prevent duplicates on next save
-            editingPersona.value = JSON.parse(JSON.stringify(newPersona));
-        } else {
-            await updatePersona(editingPersonaIndex.value, val);
-        }
-    }
-}
-
-async function handleHeaderDelete() {
-    const isPersona = currentView.value === 'view-persona-edit';
-    const title = isPersona ? (t('confirm_delete_persona') || 'Delete persona?') : (t('confirm_delete_title') || 'Delete character?');
-
-    showBottomSheet({
-        title: title,
-        items: [
-            {
-                label: t('btn_delete') || 'Delete',
-                icon: '<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="currentColor"><path d="M6 19c0 1.1.9 2 2 2h8c1.1 0 2-.9 2-2V7H6v12zM19 4h-3.5l-1-1h-5l-1 1H5v2h14V4z"/></svg>',
-                iconColor: '#ff4444',
-                isDestructive: true,
-                onClick: async () => {
-                    isDeleting.value = true; // Set flag BEFORE deletion
-                    try {
-                        if (isPersona) {
-                            await deletePersona(editingPersonaIndex.value);
-                        } else {
-                            const char = editingCharacter.value;
-                            if (char && char.id && db.deleteCharacter) {
-                                await db.deleteCharacter(char.id);
-                                await markSyncDeletedEntry('character', char.id);
-                            } else {
-                                console.error('[App] Character ID is missing or db.deleteCharacter not found');
-                            }
-                        }
-                    } catch (e) {
-                        console.error('[App] Error deleting item:', e);
-                    }
-                    closeBottomSheet();
-                    
-                    if (isPersona) {
-                        currentView.value = 'view-menu';
-                    } else {
-                        currentView.value = 'view-characters';
-                    }
-                    previousViewForEditor.value = null;
-                    // isDeleting remains true until the view changes and editor unmounts, 
-                    // effectively blocking the 'save' emit on unmount.
-                }
-            },
-            {
-                label: t('btn_cancel') || 'Cancel',
-                icon: '<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="currentColor"><path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12z"/></svg>',
-                onClick: closeBottomSheet
-            }
-        ]
-    });
-}
-
-async function closeEditor() {
-    const closingView = currentView.value;
-    const prev = previousViewForEditor.value;
-    previousViewForEditor.value = null;
-    const prevSessionId = previousSessionIdForEditor.value;
-    previousSessionIdForEditor.value = null;
-    isHeaderEditorMode.value = false;
-
-    if (closingView === 'view-character-edit') {
-        editingCharacterIndex.value = -1;
-    }
-    if (closingView === 'view-persona-edit') {
-        editingPersonaIndex.value = -1;
-    }
-
-    if (prev === 'view-chat') {
-        const isEditingChar = closingView === 'view-character-edit';
-        currentView.value = 'view-chat';
-        
-        waitForComponent(chatViewRef, async (comp) => {
-            if (activeChatCharObj.value) {
-                // If we were editing character, update the active object
-                if (isEditingChar && editingCharacter.value) {
-                    activeChatCharObj.value = { ...editingCharacter.value };
-                }
-
-                if (prevSessionId) {
-                    activeChatCharObj.value.sessionId = prevSessionId;
-                }
-
-                comp.openChat(activeChatCharObj.value, () => {
-                    currentView.value = chatPreviousView.value || 'view-dialogs';
-                });
-
-                if (shouldOpenPersonasOnReturn.value) {
-                    shouldOpenPersonasOnReturn.value = false;
-                    await nextTick();
-                    if (typeof comp.openPersonas === 'function') {
-                        comp.openPersonas();
-                    }
-                }
-            }
-        });
-    } else if (prev) {
-        currentView.value = prev;
-    } else {
-        // Fallback if previousView was not set
-        if (closingView === 'view-persona-edit') {
-            currentView.value = 'view-menu';
-        } else {
-            currentView.value = 'view-characters';
-        }
-    }
-}
-
-function finishOnboarding() {
-    localStorage.setItem('glaze_onboarding_completed', 'true');
-    isOnboarding.value = false;
-}
-
-// Full Screen Editor Logic
-const fsEditorValue = ref("");
-let fsEditorCallback = null;
-
-function openFsEditor({ value, onSave }) {
-    fsEditorValue.value = value;
-    fsEditorCallback = onSave;
-    fsEditorVisible.value = true;
-}
-
-function closeAndSaveFsEditor() {
-    if (fsEditorCallback) fsEditorCallback(fsEditorValue.value);
-    fsEditorVisible.value = false;
-}
-
-function autoSaveFsEditor(val) {
-    if (fsEditorCallback) fsEditorCallback(val);
-}
+const showLogo = computed(() => {
+    return !['view-chat', 'view-character-edit', 'view-persona-edit', 'view-theme-settings', 'view-settings',
+             'view-tools', 'view-api', 'view-presets', 'view-lorebook', 'view-regex', 'view-personas'].includes(nav.currentView.value) && !nav.isHeaderEditorMode.value;
+});
 
 const mainStyle = computed(() => {
     if (!themeState.hasBackgroundImage) return {};
@@ -480,14 +223,15 @@ const mainStyle = computed(() => {
     };
 });
 
-watch(fsEditorVisible, (val) => {
+// --- FS editor header watcher ---
+watch(editor.fsEditorVisible, (val) => {
     if (val) {
         publishAppEvent(APP_EVENTS.ui.header.setupEditor, {
             title: translations[currentLang.value]?.header_editor || 'Editor',
-            onBack: () => { fsEditorVisible.value = false; },
+            onBack: () => { editor.fsEditorVisible.value = false; },
             actions: [{
                 icon: '<svg viewBox="0 0 24 24"><path d="M9 16.17L4.83 12l-1.42 1.41L9 19 21 7l-1.41-1.41z"/></svg>',
-                onClick: closeAndSaveFsEditor
+                onClick: editor.closeAndSaveFsEditor
             }]
         });
     } else {
@@ -496,378 +240,34 @@ watch(fsEditorVisible, (val) => {
     }
 });
 
-// --- Character Editor Config ---
-const characterEditorConfig = [
-    {
-        title: 'section_basic_info',
-        fields: [
-            { key: 'name', label: 'label_name', type: 'text' },
-            { key: 'macro_name', label: 'label_char_macro_name', type: 'text', placeholder: 'placeholder_char_macro_name' },
-            { key: 'description', label: 'label_description', type: 'textarea', rows: 3, expandable: true },
-            { key: 'creator_notes', label: 'label_creator_notes', type: 'textarea', rows: 2 },
-            { key: 'tags', label: 'label_tags', type: 'tags' }
-        ]
-    },
-    {
-        title: 'section_personality',
-        fields: [
-            { key: 'personality', label: 'label_personality', type: 'textarea', rows: 4, expandable: true },
-            { key: 'scenario', label: 'label_scenario', type: 'textarea', rows: 3, expandable: true }
-        ]
-    },
-    {
-        title: 'section_dialogue',
-        fields: [
-            { key: 'first_mes', label: 'label_first_mes', type: 'greeting_list', rows: 4 },
-            { key: 'mes_example', label: 'label_mes_example', type: 'textarea', rows: 6, expandable: true }
-        ]
-    }
-];
-
-const personaEditorConfig = [
-    {
-        title: 'section_basic_info',
-        fields: [
-            { key: 'name', label: 'label_name', type: 'text' },
-            { key: 'prompt', label: 'label_description', type: 'textarea', rows: 4, expandable: true }
-        ]
-    }
-];
-
-// --- FAB Logic ---
-const fabConfig = computed(() => {
-    if (currentView.value === 'view-dialogs' && !isDesktop.value) {
-        return {
-            text: translations[currentLang.value]?.btn_new_chat || 'New Chat',
-            action: () => dialogListRef.value?.openNewChatPicker()
-        };
-    } else if (currentView.value === 'view-characters' && !isDesktop.value) {
-        return {
-            text: translations[currentLang.value]?.btn_add || 'Add',
-            action: () => characterListRef.value?.onAddCharacter()
-        };
-    }
-    return null;
-});
-
-const headerEditingIndex = computed(() => {
-    if (currentView.value === 'view-character-edit') return editingCharacterIndex.value;
-    if (currentView.value === 'view-persona-edit') return editingPersonaIndex.value;
-    return -1;
-});
-
-const showLogo = computed(() => {
-    return !['view-chat', 'view-character-edit', 'view-persona-edit', 'view-theme-settings', 'view-settings',
-             'view-tools', 'view-api', 'view-presets', 'view-lorebook', 'view-regex', 'view-personas'].includes(currentView.value) && !isHeaderEditorMode.value;
-});
-
-const updateLayoutMetrics = () => {
-    if (headerContainer.value) {
-        const h = headerContainer.value.offsetHeight;
-        document.documentElement.style.setProperty('--header-height', `${h}px`);
-    }
-    if (isDesktop.value) {
-        document.documentElement.style.setProperty('--footer-height', '0px');
-    } else if (footerContainer.value) {
-        document.documentElement.style.setProperty('--footer-height', `${footerContainer.value.offsetHeight}px`);
-    }
-};
-
-let layoutObserver = null;
-
-const onOpenCharacterEditor = (detail) => { openCharacterEditor(detail.index); };
-
-const onOpenPersonaEditor = (detail) => {
-    previousViewForEditor.value = currentView.value;
-    if (currentView.value === 'view-chat') {
-        shouldOpenPersonasOnReturn.value = true;
-    }
-    isDeleting.value = false;
-    editingPersonaIndex.value = detail.index;
-    editingPersona.value = detail.persona ? JSON.parse(JSON.stringify(detail.persona)) : { name: '', description: '', avatar: '' };
-    currentView.value = 'view-persona-edit';
-};
-
-const onNavigateTo = (detail) => { currentView.value = detail; };
-
-const onOpenOnboarding = () => { isOnboarding.value = true; };
-
-const onTriggerOpenImage = (detail) => {
-    const { src, name, description, onCloseCallback } = detail;
-    logger.debug('[App] trigger-open-image. Current Mode:', imageViewerMode.value);
-    logger.debug('[App] trigger-open-image. Always using default viewer.');
-    publishAppEvent(APP_EVENTS.nav.openImageViewer, { src, description, onCloseCallback });
-};
-
-const onOpenFsRequest = (detail) => { openFsEditor(detail); };
-
-const onOpenConnections = (detail) => {
-    const { type, id, name } = detail || {};
-    waitForComponent(connectionsSheetRef, (comp) => {
-        comp.open(type, id, name, activeChatCharObj.value);
+// --- Editor close wrapper (needs nav + editor cross-refs) ---
+function closeEditorWrapper() {
+    editor.closeEditor({
+        activeChatCharObj: nav.activeChatCharObj,
+        currentChatSessionId: nav.currentChatSessionId,
+        chatPreviousView: nav.chatPreviousView,
+        isHeaderEditorMode: nav.isHeaderEditorMode
     });
-};
-
-const onOpenItemEditor = (detail) => {
-    const { type, id } = detail;
-    if (type === 'lorebook') {
-        waitForComponent(lorebookSheetRef, (comp) => {
-            comp.openLorebook(id);
-        });
-    } else if (type === 'preset') {
-        waitForComponent(presetViewRef, (comp) => {
-            comp.openPreset(id);
-        });
-    } else if (type === 'persona') {
-        const index = allPersonas.value.findIndex(p => p.id === id);
-        if (index !== -1) {
-            const persona = allPersonas.value[index];
-            publishAppEvent(APP_EVENTS.nav.openPersonaEditor, { index, persona });
-        }
-    }
-};
-
-const onOpenLorebookEntry = (detail) => {
-    const { lorebookId, entryId } = detail;
-    if (currentView.value === 'view-chat') {
-        waitForComponent(chatViewRef, (comp) => {
-            comp.openLorebookEntry(lorebookId, entryId);
-        });
-    }
-};
-
-const onOpenBackupSheet = () => {
-    waitForComponent(backupSheetRef, (comp) => {
-        comp.open();
-    });
-};
-
-const onOpenSyncSheet = () => {
-    waitForComponent(syncSheetRef, (comp) => {
-        comp.open();
-    });
-};
-
-const onOpenConflictSheet = () => {
-    waitForComponent(conflictSheetRef, (comp) => {
-        comp.open();
-    });
-};
-
-const onOpenPresetSheet = (detail) => {
-    waitForComponent(presetViewRef, (comp) => {
-        const presetId = detail?.presetId;
-        if (presetId) {
-            comp.openPreset(presetId, true);
-        } else {
-            comp.open();
-        }
-    });
-};
-
-const onOpenApiSheet = () => {
-    waitForComponent(apiViewRef, (comp) => {
-        comp.open();
-    });
-};
-
-const onHeaderSetupEditor = () => { isHeaderEditorMode.value = true; };
-const onHeaderSetupGeneration = () => { isHeaderEditorMode.value = false; };
-const onHeaderReset = () => { isHeaderEditorMode.value = false; };
-
-const reloadSyncedData = async () => {
-    await Promise.all([
-        loadPersonas(),
-        initTheme(),
-        initLorebookState(true),
-        initPresetState(true)
-    ]);
-
-    // Re-apply synced API runtime settings so generation uses updated values immediately
-    applyApiRuntimeConfig({});
-
-    if (characterListRef.value?.loadCharacters) {
-        await characterListRef.value.loadCharacters();
-    }
-};
-
-const onSyncDataRefreshed = async () => {
-    await reloadSyncedData();
-};
-
-const handleOpenChatEvent = async (detail) => {
-    logger.debug("[App] Received open-chat event:", detail);
-    const data = detail;
-    // Handle both object (new) and string (legacy/fallback) formats
-    const charId = typeof data === 'object' ? data.charId : data;
-    const sessionId = typeof data === 'object' ? data.sessionId : null;
-    const msgId = typeof data === 'object' ? data.msgId : null;
-
-    if (!charId) return;
-
-    const chars = await db.getAll('characters');
-    const char = chars.find(c => c.id === charId);
-    if (char) {
-        if (sessionId) char.sessionId = sessionId;
-        if (msgId) char.msgId = msgId;
-        // Force back destination to dialogs — user may be on any view when tapping notification
-        currentView.value = 'view-dialogs';
-        openChatWrapper(char);
-    }
-};
-
-onMounted(async () => {
-    await migrateScToGz(); // One-time migration: sc_ -> gz_ storage keys
-
-    isOnboarding.value = localStorage.getItem('glaze_onboarding_completed') !== 'true';
-
-    initSettings();
-    await initTheme();
-    await Promise.all([
-        initLorebookState(),
-        initPresetState(),
-        loadPersonas(),
-        initSyncState()
-    ]);
-    
-    startTracking();
-    
-    if (syncProvider.value) {
-        checkSyncReadiness().then(ready => {
-            if (ready.ready) {
-                fullPull().catch(e => console.warn('[App] Background pull failed:', e));
-            }
-        });
-    }
-    
-    initRipple();
-    initThemeToggle();
-    initViewportFix();
-    initBackButton();
-    
-    initHeaderDropdown(categories, activeCategories, (viewId, itemId) => {
-        // view-dialogs is handled reactively via props to DialogList component
-        // view-characters is handled reactively via props to CharacterList component
-    });
-
-    
-    // Listen for editor events
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.openCharacterEditor, ({ detail }) => onOpenCharacterEditor(detail)));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.openPersonaEditor, ({ detail }) => onOpenPersonaEditor(detail)));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.navigateTo, ({ detail }) => onNavigateTo(detail)));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.domain.settings.languageChanged, onLanguageChanged));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.openChat, ({ detail }) => handleOpenChatEvent(detail)));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.openOnboarding, onOpenOnboarding));
-
-    const pendingData = consumePendingNotificationData();
-    if (pendingData) {
-        handleOpenChatEvent(pendingData);
-    }
-
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.triggerOpenImage, ({ detail }) => onTriggerOpenImage(detail)));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.openFsRequest, ({ detail }) => onOpenFsRequest(detail)));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.openConnections, ({ detail }) => onOpenConnections(detail)));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.openItemEditor, ({ detail }) => onOpenItemEditor(detail)));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.openLorebookEntry, ({ detail }) => onOpenLorebookEntry(detail)));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.openBackupSheet, onOpenBackupSheet));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.openSyncSheet, onOpenSyncSheet));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.openConflictSheet, onOpenConflictSheet));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.openPresetSheet, ({ detail }) => onOpenPresetSheet(detail)));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.openApiSheet, onOpenApiSheet));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.ui.header.setupEditor, onHeaderSetupEditor));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.ui.header.setupGeneration, onHeaderSetupGeneration));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.ui.header.reset, onHeaderReset));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.domain.sync.dataRefreshed, onSyncDataRefreshed));
-
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.nav.openGlossary, () => {
-        if (isDesktop.value) {
-            isGlossaryWindowOpen.value = true;
-        }
-    }));
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.ui.glossary.toggle, () => {
-        if (isDesktop.value) {
-            isGlossaryWindowOpen.value = !isGlossaryWindowOpen.value;
-        }
-    }));
-
-    // Initialize ResizeObserver for layout metrics
-    layoutObserver = new ResizeObserver(() => {
-        requestAnimationFrame(updateLayoutMetrics);
-    });
-    if (headerContainer.value) layoutObserver.observe(headerContainer.value);
-    if (footerContainer.value) layoutObserver.observe(footerContainer.value);
-    updateLayoutMetrics();
-
-    updateLanguage();
-    // AppHeader updates itself when it mounts
-    isDataLoaded.value = true;
-    
-    // Asynchronously generate missing thumbnails without blocking UI
-    generateMissingThumbnails();
-
-    checkDesktop();
-    window.addEventListener('resize', checkDesktop);
-    appEventUnsubs.push(subscribeAppEvent(APP_EVENTS.ui.glossary.headerUpdate, ({ detail }) => onGlossaryHeaderUpdate(detail)));
-
-    setTimeout(() => {
-        document.body.classList.remove('preload');
-        document.body.classList.add('app-loaded');
-    }, 100);
-
-    // Check and prompt to enable notifications on first run
-    setTimeout(checkAndRequestNotifications, 1000);
-
-    if (Capacitor.isNativePlatform()) {
-        kbListeners.push(await onKeyboardShow(() => { isKeyboardOpen.value = true; }));
-        kbListeners.push(await onKeyboardHide(() => { isKeyboardOpen.value = false; }));
-    }
-});
-
-onBeforeUnmount(() => {
-    if (layoutObserver) layoutObserver.disconnect();
-    appEventUnsubs.forEach(unsub => unsub());
-    appEventUnsubs.length = 0;
-    stopGlossaryDrag();
-    kbListeners.forEach(l => l.remove());
-    window.removeEventListener('resize', checkDesktop);
-});
-
-watch(currentView, () => {
-    isHeaderEditorMode.value = false;
-});
-
-watch(isDesktop, () => {
-    requestAnimationFrame(updateLayoutMetrics);
-});
-
-watch(forceMobileLayout, () => {
-    checkDesktop();
-});
-
-watch(currentView, (newVal, oldVal) => {
-    if (isDesktop.value && oldVal && !menuViews.includes(oldVal) && menuViews.includes(newVal)) {
-        desktopPreMenuView.value = oldVal;
-    }
-});
+}
 </script>
 
 <template>
-  <div class="app-layout" :class="{ 'desktop-mode': isDesktop }" :style="mainStyle">
+  <div class="app-layout" :class="{ 'desktop-mode': nav.isDesktop.value }" :style="mainStyle">
     <!-- Onboarding Overlay -->
     <Transition name="fade">
-        <OnboardingView v-if="isOnboarding" @finish="finishOnboarding" />
+        <OnboardingView v-if="nav.isOnboarding.value" @finish="nav.finishOnboarding" />
     </Transition>
 
     <!-- Header Component -->
     <div class="header-container" ref="headerContainer" :style="{ zIndex: headerZIndex }">
         <AppHeader
             ref="headerRef"
-            :current-view="currentView"
+            :current-view="nav.currentView.value"
             :categories="categories"
-            :editing-index="headerEditingIndex"
-            @action-save="handleHeaderSave"
-            @action-delete="handleHeaderDelete"
-            @action-close="closeEditor"
+            :editing-index="editor.headerEditingIndex.value"
+            @action-save="editor.handleHeaderSave"
+            @action-delete="editor.handleHeaderDelete"
+            @action-close="closeEditorWrapper"
         />
     </div>
 
@@ -876,132 +276,132 @@ watch(currentView, (newVal, oldVal) => {
 
       <!-- Desktop Left Sidebar: always-visible chat list + nav -->
       <DesktopLeftSidebar 
-          v-if="isDesktop"
-          :current-view="currentView"
+          v-if="nav.isDesktop.value"
+          :current-view="nav.currentView.value"
           :active-categories="activeCategories"
-          :is-desktop-floating="isDesktopFloating"
-          :is-glossary-open="isGlossaryWindowOpen"
-          @update:current-view="currentView = $event"
-          @open-chat="openChatWrapper"
+          :is-desktop-floating="nav.isDesktopFloating.value"
+          :is-glossary-open="glossary.isGlossaryWindowOpen.value"
+          @update:current-view="nav.currentView.value = $event"
+          @open-chat="openChatFromTemplate"
       />
 
       <!-- Main Content Area -->
-      <main id="main-container" v-if="isDataLoaded" :style="{ zIndex: mainZIndex }" :class="{ 'keyboard-open': isKeyboardOpen && effectiveMainView !== 'view-chat', 'chat-view-main': effectiveMainView === 'view-chat' }">
+      <main id="main-container" v-if="isDataLoaded" :style="{ zIndex: mainZIndex }" :class="{ 'keyboard-open': isKeyboardOpen && nav.effectiveMainView.value !== 'view-chat', 'chat-view-main': nav.effectiveMainView.value === 'view-chat' }">
 
         <Transition name="fade">
             <!-- VIEW 1: DIALOGS -->
-            <div id="view-dialogs" class="view active-view" v-if="effectiveMainView === 'view-dialogs' && !isDesktop">
+            <div id="view-dialogs" class="view active-view" v-if="nav.effectiveMainView.value === 'view-dialogs' && !nav.isDesktop.value">
                 <DialogList
                     ref="dialogListRef"
                     :active-category="activeCategories['view-dialogs']"
-                    @open-chat="openChatWrapper"
+                    @open-chat="openChatFromTemplate"
                 />
             </div>
 
             <!-- VIEW 2: CHARACTERS -->
-            <div id="view-characters" class="view active-view" v-else-if="effectiveMainView === 'view-characters'">
+            <div id="view-characters" class="view active-view" v-else-if="nav.effectiveMainView.value === 'view-characters'">
                 <CharacterList
                     ref="characterListRef"
                     :active-category="activeCategories['view-characters']"
-                    @open-chat="openChatWrapper"
+                    @open-chat="openChatFromTemplate"
                 />
             </div>
 
             <!-- VIEW 3: MENU -->
             <MenuView
                 class="view active-view view-gray-bg"
-                v-else-if="effectiveMainView === 'view-menu' && !isDesktopFloating"
+                v-else-if="nav.effectiveMainView.value === 'view-menu' && !nav.isDesktopFloating.value"
             />
 
             <!-- VIEW: GLOSSARY -->
             <GlossaryView
                 class="view active-view view-gray-bg"
-                v-else-if="effectiveMainView === 'view-glossary' && !isDesktopFloating"
+                v-else-if="nav.effectiveMainView.value === 'view-glossary' && !nav.isDesktopFloating.value"
                 :view-mode="true"
             />
 
             <!-- VIEW: THEME SETTINGS -->
             <ThemeSettingsView
                 class="view active-view view-gray-bg"
-                v-else-if="effectiveMainView === 'view-theme-settings' && !isDesktopFloating"
+                v-else-if="nav.effectiveMainView.value === 'view-theme-settings' && !nav.isDesktopFloating.value"
             />
 
             <!-- VIEW: SETTINGS -->
             <SettingsView
                 class="view active-view view-gray-bg"
-                v-else-if="effectiveMainView === 'view-settings' && !isDesktopFloating"
+                v-else-if="nav.effectiveMainView.value === 'view-settings' && !nav.isDesktopFloating.value"
             />
 
             <!-- VIEW: TOOLS HUB -->
             <ToolsView
                 class="view active-view view-gray-bg"
-                v-else-if="effectiveMainView === 'view-tools'"
+                v-else-if="nav.effectiveMainView.value === 'view-tools'"
             />
 
             <!-- VIEW: API (fullscreen) -->
             <ApiView
                 class="view active-view view-gray-bg"
-                v-else-if="effectiveMainView === 'view-api'"
+                v-else-if="nav.effectiveMainView.value === 'view-api'"
                 :view-mode="true"
             />
 
             <!-- VIEW: PRESETS (fullscreen) -->
             <PresetView
                 class="view active-view view-gray-bg"
-                v-else-if="effectiveMainView === 'view-presets'"
+                v-else-if="nav.effectiveMainView.value === 'view-presets'"
                 :view-mode="true"
             />
 
             <!-- VIEW: LOREBOOK (fullscreen) -->
             <LorebookSheet
                 class="view active-view view-gray-bg"
-                v-else-if="effectiveMainView === 'view-lorebook'"
+                v-else-if="nav.effectiveMainView.value === 'view-lorebook'"
                 :view-mode="true"
             />
 
             <!-- VIEW: REGEX (fullscreen) -->
             <RegexSheet
                 class="view active-view view-gray-bg"
-                v-else-if="effectiveMainView === 'view-regex'"
+                v-else-if="nav.effectiveMainView.value === 'view-regex'"
                 :view-mode="true"
             />
 
             <!-- VIEW: PERSONAS (fullscreen) -->
             <PersonasView
                 class="view active-view view-gray-bg"
-                v-else-if="effectiveMainView === 'view-personas'"
+                v-else-if="nav.effectiveMainView.value === 'view-personas'"
                 :view-mode="true"
             />
 
             <!-- VIEW 5: CHAT -->
-            <ChatView class="view active-view" v-else-if="effectiveMainView === 'view-chat'" ref="chatViewRef" />
+            <ChatView class="view active-view" v-else-if="nav.effectiveMainView.value === 'view-chat'" ref="chatViewRef" />
 
             <!-- VIEW 6: CHARACTER / PERSONA EDITOR -->
             <Editor
                 class="view active-view"
-                v-else-if="effectiveMainView === 'view-character-edit' || effectiveMainView === 'view-persona-edit'"
-                :model-value="effectiveMainView === 'view-character-edit' ? (editingCharacter || {}) : (editingPersona || {})"
-                :config="effectiveMainView === 'view-character-edit' ? characterEditorConfig : personaEditorConfig"
+                v-else-if="nav.effectiveMainView.value === 'view-character-edit' || nav.effectiveMainView.value === 'view-persona-edit'"
+                :model-value="nav.effectiveMainView.value === 'view-character-edit' ? (editor.editingCharacter.value || {}) : (editor.editingPersona.value || {})"
+                :config="nav.effectiveMainView.value === 'view-character-edit' ? characterEditorConfig : personaEditorConfig"
                 :show-avatar="true"
-                @update:model-value="(val) => effectiveMainView === 'view-character-edit' ? editingCharacter = val : editingPersona = val"
-                @save="handleEditorAutoSave"
-                @close="closeEditor"
-                @open-fs="openFsEditor"
+                @update:model-value="(val) => nav.effectiveMainView.value === 'view-character-edit' ? editor.editingCharacter.value = val : editor.editingPersona.value = val"
+                @save="editor.handleEditorAutoSave"
+                @close="closeEditorWrapper"
+                @open-fs="editor.openFsEditor"
             />
         </Transition>
       </main>
 
       <!-- Desktop: Floating menu overlay -->
       <Transition name="fade">
-          <div v-if="isDesktopFloating" class="desktop-float-overlay" @click.self="currentView = desktopPreMenuView">
+          <div v-if="nav.isDesktopFloating.value" class="desktop-float-overlay" @click.self="nav.currentView.value = nav.desktopPreMenuView.value">
               <div class="desktop-float-panel">
-                  <button class="desktop-float-close" @click="currentView = desktopPreMenuView">
+                  <button class="desktop-float-close" @click="nav.currentView.value = nav.desktopPreMenuView.value">
                       <svg viewBox="0 0 24 24"><path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12z"/></svg>
                   </button>
-                  <MenuView v-if="currentView === 'view-menu'" class="view-gray-bg" />
+                  <MenuView v-if="nav.currentView.value === 'view-menu'" class="view-gray-bg" />
                   <!-- glossary has its own corner popup, not shown here -->
-                  <ThemeSettingsView v-else-if="currentView === 'view-theme-settings'" class="view-gray-bg" />
-                  <SettingsView v-else-if="currentView === 'view-settings'" class="view-gray-bg" />
+                  <ThemeSettingsView v-else-if="nav.currentView.value === 'view-theme-settings'" class="view-gray-bg" />
+                  <SettingsView v-else-if="nav.currentView.value === 'view-settings'" class="view-gray-bg" />
               </div>
           </div>
 
@@ -1009,13 +409,13 @@ watch(currentView, (newVal, oldVal) => {
 
       <!-- Desktop: Glossary corner popup -->
       <Transition name="glossary-popup">
-          <div v-if="isDesktopGlossary" class="desktop-glossary-popup" :style="glossaryStyle">
-              <div class="desktop-glossary-popup-header" @mousedown="startGlossaryDrag">
-                  <button v-if="glossaryCanGoBack" class="desktop-glossary-popup-back" @click="onGlossaryBack">
+          <div v-if="glossary.isDesktopGlossary.value" class="desktop-glossary-popup" :style="glossary.glossaryStyle.value">
+              <div class="desktop-glossary-popup-header" @mousedown="glossary.startGlossaryDrag">
+                  <button v-if="glossary.glossaryCanGoBack.value" class="desktop-glossary-popup-back" @click="glossary.onGlossaryBack">
                       <svg viewBox="0 0 24 24"><path d="M20 11H7.83l5.59-5.59L12 4l-8 8 8 8 1.41-1.41L7.83 13H20v-2z"/></svg>
                   </button>
-                  <span class="desktop-glossary-popup-title" :class="{ 'pl-3': !glossaryCanGoBack }">{{ glossaryPopupTitle }}</span>
-                  <button class="desktop-glossary-popup-close" @click="isGlossaryWindowOpen = false">
+                  <span class="desktop-glossary-popup-title" :class="{ 'pl-3': !glossary.glossaryCanGoBack.value }">{{ glossary.glossaryPopupTitle.value }}</span>
+                  <button class="desktop-glossary-popup-close" @click="glossary.isGlossaryWindowOpen.value = false">
                       <svg viewBox="0 0 24 24"><path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12z"/></svg>
                   </button>
               </div>
@@ -1025,11 +425,11 @@ watch(currentView, (newVal, oldVal) => {
 
       <!-- Desktop Right Sidebar: MagicDrawer panel -->
       <DesktopRightSidebar
-          v-if="isDesktop"
+          v-if="nav.isDesktop.value"
           :bottom-sheet-state="bottomSheetState"
           :right-sidebar-state="sidebarState"
-          :active-chat-char-obj="activeChatCharObj"
-          :current-view="currentView"
+          :active-chat-char-obj="nav.activeChatCharObj.value"
+          :current-view="nav.currentView.value"
           @close-bottom-sheet="closeBottomSheet"
           @magic-notes="chatViewRef?.openAuthorsNoteSheet()"
           @magic-context="chatViewRef?.openContextSheet()"
@@ -1044,7 +444,7 @@ watch(currentView, (newVal, oldVal) => {
           @magic-memory-books="chatViewRef?.openMemoryBooksSheet()"
           @magic-regex="chatViewRef?.openRegexSheet()"
           @magic-image-gen="chatViewRef?.openImageGenSheet()"
-          @magic-glossary="isGlossaryWindowOpen = true"
+          @magic-glossary="glossary.isGlossaryWindowOpen.value = true"
       />
 
       <!-- Floating Action Button: Positioned relative to grid on Desktop -->
@@ -1060,8 +460,8 @@ watch(currentView, (newVal, oldVal) => {
     </div>
 
     <!-- Bottom Navigation Bar (mobile only — desktop uses left sidebar) -->
-    <div v-if="!isDesktop" class="footer-container" ref="footerContainer">
-        <BottomNavigation v-model:current-view="currentView" />
+    <div v-if="!nav.isDesktop.value" class="footer-container" ref="footerContainer">
+        <BottomNavigation v-model:current-view="nav.currentView.value" />
     </div>
 
     <!-- Global Bottom Sheet -->
@@ -1083,9 +483,6 @@ watch(currentView, (newVal, oldVal) => {
     <!-- Global Desktop Dropdown (replaces bottom sheets for simple selects on PC) -->
     <DesktopDropdown />
 
-    <!-- Holo Cards Viewer -->
-    <!-- <HoloCardViewer /> -->
-
     <!-- Standard Image Viewer -->
     <ImageViewer />
 
@@ -1095,10 +492,10 @@ watch(currentView, (newVal, oldVal) => {
 
   <!-- Full Screen Editor (Managed by App.vue now) -->
   <FullScreenEditor 
-      :visible="fsEditorVisible"
-      v-model="fsEditorValue"
-      @save="autoSaveFsEditor"
-      @close="fsEditorVisible = false"
+      :visible="editor.fsEditorVisible.value"
+      v-model="editor.fsEditorValue.value"
+      @save="editor.autoSaveFsEditor"
+      @close="editor.fsEditorVisible.value = false"
   />
 
   <ConnectionsSheet ref="connectionsSheetRef" />
