+++ src/core/llm/transport/streamingSse.js
@@ -8,7 +8,6 @@ export async function consumeStreamingSseResponse({
     controller,
     streamTimeout,
     throwIfAborted,
-    requestReasoning,
     streamAccumulator,
     onUpdate
 }) {
